ci: fix CircleCI
the `release` job depended on the `node-*` jobs, but the `node-*` jobs
did not declare a `tags` filter, so they were not run on tags.

Per the [CircleCI documentation][1]:

> Additionally, if a job requires any other jobs (directly or
indirectly), you must specify tag filters for those jobs.

This commit updates the `node-*` jobs to declare a `tags` filter so they
always run on every job. It also updates the CircleCI config to update
the `node` jobs to use CircleCI's matrix feature, so we don't have to
keep track of configurations between node versions. This will make
updating CircleCI easier when we need to update Node.js based on the
Node.js LTS release schedule

[1]: https://circleci.com/docs/2.0/configuration-reference/#tags