Add missing topics to package.json (#468)
* add "compute collaborator" topics to package.json

* fix: update topic description

* add "resume" sub-topic

I know that the functions plugin doesn't contain any `sf env resume` commands.  I'm working around a race-condition with plugin-env that DOES have `sf env resume` commands.  The plugin-functions package.json is "winning" and so the help for the env commands is incorrect.

* removing this now-unneeded addition 

Turns out that version 1.5.1 of plugin-env fixes this issue, so there's no need to add this unneeded topic to the package.json of `plugins-functions`.  So I'm removing it!  But leaving the other addition, because that IS needed.