chore(tests): Clean up `run:function:start:local` tests
Previously:
- The tests in many cases didn't test what they said they did
  (for example, there being a `package.json` in the fixture plays
   no part on how the local runner is invoked itself, and some tests
   said they were testing the long form arguments when they were
   testing the short form instead).
- The template fixtures were out of date and not actually used.
- There was no test coverage of:
   - the default path (ie: when `--path` is not specified)
   - the port and debug-port arguments

I believe much of this is leftover from (a) the initial dockerless
transition plus the split out to `sf-functions-core`.

If in the future we decide to add end to end tests to this repository
(and not just `sf-functions-core`), then we can add up to date
functions fixtures back at that point.

GUS-W-11643273.