fix: show the HTTP body for functions non HTTP-200 responses
Now when using `sf run function`, if the function server responds
with a non-HTTP-200 status code, the body of the response is printed
too, rather than only the status code.

This makes debugging the functions runtime (and any issues
with `plugin-functions` or `sf-functions-core`, such as
https://github.com/heroku/sf-functions-core/pull/81) much
easier.

This also means the `plugin-functions` behaviour now matches
that of `salesforcedx-vscode`, which already outputs the response
body for errors:
https://github.com/forcedotcom/salesforcedx-vscode/blob/e391707d0225de338f4f8e0e79e0a5cdd6b51f11/packages/salesforcedx-vscode-core/src/commands/functions/forceFunctionInvoke.ts#L61

GUS-W-11725595.