fix(deps): Switch back to non-shortened @heroku/functions-core dependency name
In salesforcecli/plugin-functions#403 and heroku/sf-functions-core#22 various
measures were taken to try and shorten the maximum on-disk filepath of files
in the two packages, so as to prevent the CI failures seen in:
https://app.circleci.com/pipelines/github/salesforcecli/plugin-functions/2415/workflows/510a1bac-1a64-4a1b-a5a5-e765100b08d2/jobs/8514

One of these measure was to use a package alias of `@hk/functions-core` for the
`@heroku/functions-core` package, so the on-disk directory path was 4 characters
shorter.

However, this change stops Dependabot from working, and looking at recent tarfile
verification logs in CI, it appears we might not need to save these four characters any
more - since none of the salesforce-core dep paths appear in the "close to the limit"
warnings.

As such, we can undo this aliasing, to fix Dependabot for this package.

@W-12466399@